requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
httpx>=0.27.0
pydantic>=2.7.0
openai>=1.52.0
pytest>=8.2.0
python-dotenv>=1.0.1
